Began work towards drawing tasks from objects,  gpt-j API
Currently, tasks are drawn from an array called taskNames, which is created from an array of objects called tasks. Working towards skipping taskNames step so drawn tasks can be identified with ID, priority, etc. Also, successfully implemented a button that tells the users a generated joke using gpt-j (this is likely temporary and will be used for a different project).